New manifest for L25.10RC1
This is a baseline manual merges to provide a common working set
of patches for developers and to provide a base release from which
some testing can begin.

It only has the kernel patches -- audio patches from Misael and
the bridge patches do not apply.  The merge conflicts are being worked on.