Updating the fastboot procedure for NAND flashing
We should always boot out of the latest u-boot before we flash the
new software

This script will first erase all the NAND partitions, ensuring everything
is in a good state prior to flashing. Once that is done, we will flash the
the latest u-boot and XLoader. The user is then expected to "Power Cycle"
to ensure that they boot with latest u-boot. Doing this will ensure that
the fastboot client will know about changed partition table etc ...

Change-Id: Ide9fda06deeedd481bcf7d29ac75b356057ba230