Updating the fastboot procedure for eMMC flashing
We should always boot out of the latest u-boot before we flash the
new software

This script will first erase all the eMMC partitions, ensuring everything
is in a good state prior to flashing. Once that is done, we will flash the
the latest mbr, u-boot and XLoader. The user is then expected to "Power Cycle"
to ensure that they boot with latest u-boot. Doing this will ensure that
the fastboot client will know about changed partition table etc ...

Change-Id: Ie5a089c2531d286f91e9a484565011867e0572b5
Signed-off-by: Akash Choudhari <akashc@ti.com>