OMAP4: Update fastboot scripts to use reboot-bootloader command
* On blaze - keeping any key pressed and rebooting the system puts the
uboot automatically in fastboot mode.
   This eliminates to manually type fastboot on uboot prompt.

* ./fastboot reboot-bootloader
Is a new command implemented that allows system in fastboot mode to
reboot and load new uboot from emmc and enter fastboot again.

* ./fastboot reboot
This will reset the system , reload the uboot and start the boot process

All these commands are implemented to allow more scripting via fastboot.
You will never need a usb serial console to do fastboot now on.

Now on, prepare_fastboot.sh should not be executed. The script is
present to maintain compitability with froyo

Change-Id: Ib437b9a3e0cb281e92e8067792df944e25ca5156
Signed-off-by: Akash Choudhari <akashc@ti.com>
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>