Modified parser to account for new google-added seperators
Google has added a new field in the manifest that uses colons to
seperate data.  This script, as it is not using a real XML parser, was
also arbitrarily using colons as seperators.  For the short term, I've
modified the internal seperator to be a caret.

I've also removed a lot of whitespace, and added "--no-pager" to the
git command to avoid being piped through less.

Change-Id: I3f8c0abe58c09585dcdc9fa03dff0d6a3fa5366a
Signed-off-by: James W. Mills <jameswmills@ti.com>